yield from by the book
In Issues no. 691 and 692, pyos describes some differences in behavior
between our current `yield-from` backport macro and the actual Python
3.3+ `yield from` expression: specifically, that `yield-from` yields (if
you will) a SyntaxError ("'return' with argument inside generator") when
an expression is in tail position, and that `yield-from` fails to inject
exceptions into the subgenerator being yielded from.

This commit _begins_ to attempt to fix both issues by reimplementing
`yield-from` in a way that tries to mirror as closely as possible the
implementation given in the "Formal Semantics" section of PEP
380 ("Syntax for Delegating to a Subgenerator", in which the `yield
from` expression is defined). In particular, the "'return' with
argument" error seems to have been due to Hy translating the last
subexpression in a `do` form into the value returned by a Python
function, which won't work with pre-PEP 380 generators; instead, we must
explicitly raise StopIteration with an argument.

However, there are some serious limitations to note!—the implementation
in this commit is lacking in macro hygiene, and has been observed to
break (on the "'return' with argument" error again) in at least one
situation where the analogous Python 3.3+ code is fine.